Initial GUI and GUI Controls
Created a very basic GUI window with a send command button and 4 commands to be able to send. When a command is selected and the send command button pressed, a message box will appear displaying the sent command. If no command is selected a message box displaying an error will appear.