Additional GUI Windows and Prototype Networking Code
Created dedicated windows for the client/server configuration. They contain information relevant to each operating mode.

Added in most of the networking code. It currently does NOT work and is designed to transfer files not instructions. Once file transfer is working a small conversion is needed to set it up to handle bit streams and everything moved to their appropriate classes. When everything is in the correct location full GUI development can begin.